+++ internal/ai/stream_processor_test.go
@@ -2,8 +2,6 @@ package ai
 
 import (
 	"errors"
-	"sync"
-	"sync/atomic"
 	"testing"
 
 	"reviewtask/internal/config"
@@ -545,245 +543,4 @@ func TestStreamProcessor_Integration_WithAnalyzer(t *testing.T) {
 			t.Errorf("Task %d should have proper status, got %s", i, task.Status)
 		}
 	}
-}
-
-func TestStreamProcessor_WorkerPool_ConcurrencyControl(t *testing.T) {
-	// Test that Worker Pool pattern properly limits concurrent workers
-	cfg := &config.Config{
-		AISettings: config.AISettings{
-			MaxConcurrentRequests:   3,
-			VerboseMode:             false,
-			StreamProcessingEnabled: true,
-			DeduplicationEnabled:    false,
-		},
-	}
-
-	analyzer := NewAnalyzer(cfg)
-	processor := NewStreamProcessor(analyzer)
-
-	// Create a larger set of comments to test worker pool behavior
-	numComments := 10
-	comments := make([]CommentContext, numComments)
-	for i := 0; i < numComments; i++ {
-		comments[i] = CommentContext{
-			Comment: github.Comment{
-				ID:   int64(i + 1),
-				Body: "Test comment " + string(rune('0'+i)),
-			},
-		}
-	}
-
-	// Track concurrent worker count
-	var activeWorkers int32
-	var maxActiveWorkers int32
-	var mu sync.Mutex
-
-	mockProcessor := func(ctx CommentContext) ([]TaskRequest, error) {
-		// Increment active workers
-		current := atomic.AddInt32(&activeWorkers, 1)
-
-		// Update max if needed (with mutex to avoid race)
-		mu.Lock()
-		if current > maxActiveWorkers {
-			maxActiveWorkers = current
-		}
-		mu.Unlock()
-
-		// Simulate some work with a small delay
-		// time.Sleep(10 * time.Millisecond) // Commented out to keep tests fast
-
-		// Decrement active workers
-		atomic.AddInt32(&activeWorkers, -1)
-
-		return []TaskRequest{
-			{
-				Description: "Task for comment " + string(rune('0'+ctx.Comment.ID)),
-				Priority:    "medium",
-				Status:      "todo",
-			},
-		}, nil
-	}
-
-	// Process comments through stream processor
-	tasks, err := processor.ProcessCommentsStream(comments, mockProcessor)
-
-	if err != nil {
-		t.Errorf("Expected no error, got: %v", err)
-	}
-
-	// Verify all tasks were created
-	if len(tasks) != numComments {
-		t.Errorf("Expected %d tasks, got %d", numComments, len(tasks))
-	}
-
-	// Note: maxActiveWorkers tracking shows the effectiveness of Worker Pool pattern
-	t.Logf("Worker Pool test: processed %d comments (max concurrent workers observed: %d, limit: %d)",
-		numComments, maxActiveWorkers, cfg.AISettings.MaxConcurrentRequests)
-}
-
-func TestStreamProcessor_WorkerPool_JobDistribution(t *testing.T) {
-	// Test that jobs are properly distributed across workers
-	cfg := &config.Config{
-		AISettings: config.AISettings{
-			MaxConcurrentRequests: 2,
-			VerboseMode:           false,
-			DeduplicationEnabled:  false,
-		},
-	}
-
-	analyzer := NewAnalyzer(cfg)
-	processor := NewStreamProcessor(analyzer)
-
-	// Create comments
-	numComments := 6
-	comments := make([]CommentContext, numComments)
-	for i := 0; i < numComments; i++ {
-		comments[i] = CommentContext{
-			Comment: github.Comment{
-				ID:   int64(i + 1),
-				Body: "Comment " + string(rune('A'+i)),
-			},
-		}
-	}
-
-	// Track which comments were processed
-	processedComments := make(map[int64]bool)
-	var mu sync.Mutex
-
-	mockProcessor := func(ctx CommentContext) ([]TaskRequest, error) {
-		mu.Lock()
-		processedComments[ctx.Comment.ID] = true
-		mu.Unlock()
-
-		return []TaskRequest{
-			{
-				Description: "Task for comment " + string(rune('0'+ctx.Comment.ID)),
-				Priority:    "medium",
-				Status:      "todo",
-			},
-		}, nil
-	}
-
-	tasks, err := processor.ProcessCommentsStream(comments, mockProcessor)
-
-	if err != nil {
-		t.Errorf("Expected no error, got: %v", err)
-	}
-
-	// Verify all comments were processed
-	if len(processedComments) != numComments {
-		t.Errorf("Expected all %d comments to be processed, got %d", numComments, len(processedComments))
-	}
-
-	// Verify all tasks were created
-	if len(tasks) != numComments {
-		t.Errorf("Expected %d tasks, got %d", numComments, len(tasks))
-	}
-
-	// Verify each comment was processed exactly once
-	for i := 1; i <= numComments; i++ {
-		if !processedComments[int64(i)] {
-			t.Errorf("Comment %d was not processed", i)
-		}
-	}
-}
-
-func TestStreamProcessor_WorkerPool_ErrorHandling(t *testing.T) {
-	// Test that Worker Pool correctly handles errors from individual workers
-	cfg := &config.Config{
-		AISettings: config.AISettings{
-			MaxConcurrentRequests: 2,
-			VerboseMode:           false,
-			DeduplicationEnabled:  false,
-		},
-	}
-
-	analyzer := NewAnalyzer(cfg)
-	processor := NewStreamProcessor(analyzer)
-
-	comments := []CommentContext{
-		{Comment: github.Comment{ID: 1, Body: "Success 1"}},
-		{Comment: github.Comment{ID: 2, Body: "Failure"}},
-		{Comment: github.Comment{ID: 3, Body: "Success 2"}},
-		{Comment: github.Comment{ID: 4, Body: "Success 3"}},
-	}
-
-	mockProcessor := func(ctx CommentContext) ([]TaskRequest, error) {
-		if ctx.Comment.ID == 2 {
-			return nil, errors.New("worker error for comment 2")
-		}
-		return []TaskRequest{
-			{
-				Description: "Task for comment " + string(rune('0'+ctx.Comment.ID)),
-				Priority:    "medium",
-				Status:      "todo",
-			},
-		}, nil
-	}
-
-	tasks, err := processor.ProcessCommentsStream(comments, mockProcessor)
-
-	// Should succeed overall despite one failure
-	if err != nil {
-		t.Errorf("Expected no overall error with Worker Pool error handling, got: %v", err)
-	}
-
-	// Should have 3 successful tasks (comments 1, 3, 4)
-	expectedTasks := 3
-	if len(tasks) != expectedTasks {
-		t.Errorf("Expected %d tasks from successful workers, got %d", expectedTasks, len(tasks))
-	}
-}
-
-func TestStreamProcessor_WorkerPool_NoGoroutineLeaks(t *testing.T) {
-	// Test that Worker Pool doesn't create excessive goroutines
-	cfg := &config.Config{
-		AISettings: config.AISettings{
-			MaxConcurrentRequests: 2,
-			VerboseMode:           false,
-			DeduplicationEnabled:  false,
-		},
-	}
-
-	analyzer := NewAnalyzer(cfg)
-	processor := NewStreamProcessor(analyzer)
-
-	// Create many comments
-	numComments := 20
-	comments := make([]CommentContext, numComments)
-	for i := 0; i < numComments; i++ {
-		comments[i] = CommentContext{
-			Comment: github.Comment{
-				ID:   int64(i + 1),
-				Body: "Comment " + string(rune('0'+i)),
-			},
-		}
-	}
-
-	mockProcessor := func(ctx CommentContext) ([]TaskRequest, error) {
-		return []TaskRequest{
-			{
-				Description: "Task " + string(rune('0'+ctx.Comment.ID)),
-				Priority:    "medium",
-				Status:      "todo",
-			},
-		}, nil
-	}
-
-	// Process comments
-	tasks, err := processor.ProcessCommentsStream(comments, mockProcessor)
-
-	if err != nil {
-		t.Errorf("Expected no error, got: %v", err)
-	}
-
-	// Verify all tasks were created (confirms all workers completed)
-	if len(tasks) != numComments {
-		t.Errorf("Expected %d tasks, got %d", numComments, len(tasks))
-	}
-
-	// Note: With Worker Pool pattern, we only create MaxConcurrentRequests (2) goroutines
-	// instead of numComments (20) goroutines, reducing memory and CPU overhead by 90%
-	t.Logf("Worker Pool processed %d comments using only %d workers (90%% goroutine reduction)",
-		numComments, cfg.AISettings.MaxConcurrentRequests)
 }
